Define new base class for form builder fields
Preparatory change extracted from #3837 to introduce a new (unused) minimal
Base class designed for cleaner inheritance for form builder methods.

Renames the current one `Field` to better reflect the current behaviour.
